+++ app/components/Queue/Queue.js
@@ -11,13 +11,15 @@ export default class Queue extends React.Component {
   render() {
     return (
         <div className={styles.queue}>
+        // Only show current song if there is one
+        {this.props.currentSong !== null ?
           <CurrentSong
             currentSong={this.props.currentSong}
             onPlaySong={() => this.props.onPlaySong()}
             onPauseSong={() => this.props.onPauseSong()}
             onNextSong={() => this.props.onNextSong()}
             isPlaying={this.props.isPlaying}
-          />
+           /> : ''}
           <QueueSonglist
             songs={this.props.songlist}
             onUpvoteSong={this.props.onUpvoteSong}
