+++ app/main.js
@@ -5,8 +5,8 @@ import io from 'socket.io-client';
 import reducer from './reducer';
 import {createStore, applyMiddleware} from 'redux';
 import {Provider} from 'react-redux';
-import setState from './actions/setstate';
 import remoteActionMiddleware from './remote_action_middleware';
+import setState from '../common/actions/setstate';
 
 const socket = io(`${location.protocol}//${location.hostname}:8090/partyq`);
 
@@ -20,19 +20,6 @@ socket.on('state', state => {
   console.log('STATE RECEIVED', state);
   store.dispatch(setState(state));
 });
-
-store.dispatch({type: 'ADD_SONG', url: 'https://www.youtube.com/watch?v=nfWlot6h_JM'});
-// store.dispatch({type: 'ADD_SONG', url: 'https://www.youtube.com/watch?v=4d2lGAP5xvQ'});
-// store.dispatch({type: 'ADD_SONG', url: 'https://www.youtube.com/watch?v=PhRa3REdozw'});
-// console.log(store.getState());
-// store.dispatch({type: 'UPVOTE_SONG', index: 1});
-// console.log(store.getState());
-// store.dispatch({type: 'PLAY_SONG'});
-// console.log(store.getState());
-// store.dispatch({type: 'PAUSE_SONG'});
-// console.log(store.getState());
-// store.dispatch({type: 'NEXT_SONG'});
-// console.log(store.getState());
 
 ReactDOM.render(
     <Provider store={store}>
