+++ app/remote_action_middleware.js
@@ -1,7 +1,10 @@
+import {SET_STATE} from '../common/constants/ActionTypes';
+
 export default socket => store => next => action => {
-  if (action !== 'SET_STATE') {
+  console.log('ACTION', action);
+  if (action.type !== SET_STATE) {
     socket.emit('action', action);
-    return next(action);
   }
+
   return next(action);
 };
