+++ common/actions/currentSong.js
